Added basic style elements, made homepage a php file, added the headed
To access php filles with localhost I made a new clone of the repository in the href folder in XAMPP.
The header is just something basic that Ben E may have to edit (or make a new one). The idea is that all documents can include it to keep the website consistent.