Database, Login, Logout, Check if Logged in, Delete Notes
The Database had some changes made to it nothing major but just mostly changes to attribute sizes and requirements and IMPORTANTLY some INSERT INTO queries at the bottom. These queries are for some employees and machines. It includes all machines required to import the csv. However I have only been able to insert all values but timestamps because they are in a different format.

The login and logout pages now have functionality, they aren't fully finished but as long as you don't try to brake it it should be fine.

Every page will have to require_once a php file called loggedin.inc.php.
The exact line of code is...
 <?php require_once "../inc/loggedin.inc.php"; ?>
You may or may not need the php tags

If anyone else needs to access the users employee id or role first make sure that you have used the line of code above then use $_SESSION['employeeID'] and $_SESSION['role']

Notes can now be deleted and notes are made by the logged in user and only visible to that user.